Treat any terminal status entry as a finished bot since terminal states are absorbing
